Apply admin health check on prerenderer

+++ web/scripts/prerender.js
@@ -1,18 +1,79 @@
 const path = require("path");
 const fs = require("fs-extra");
+const { default: fetch } = require("node-fetch");
 const { default: nextBuild } = require("next/dist/build");
 const { default: nextExport } = require("next/dist/export");
 const { trace } = require("next/dist/trace");
-const { STATIC_WEB_ROOT } = require("dotenv").config().parsed;
+const { STATIC_WEB_ROOT, NEXT_PUBLIC_ADMIN_API_ROOT } =
+  require("dotenv").config().parsed;
 
 const PROJECT_ROOT = path.resolve(__dirname, "../");
 const OUT_DIR = path.resolve(PROJECT_ROOT, "out");
 
+const ADMIN_HEALTHCHECK_INTERVAL = 10000;
+const ADMIN_HEALTHCHECK_MAX_TRIES = 60;
+
 // Source: next/dist/cli/next-export
 const span = trace("next-export-cli");
 
+const adminHealthCheck = () =>
+  new Promise(async (resolve) => {
+    let tries = ADMIN_HEALTHCHECK_MAX_TRIES;
+
+    while (tries) {
+      try {
+        const response = await fetch(NEXT_PUBLIC_ADMIN_API_ROOT, {
+          method: "post",
+          headers: {
+            "content-type": "application/json",
+          },
+          body: JSON.stringify({
+            query: `
+              query {
+                products(limit: 1) {
+                  id
+                }
+              }
+            `,
+          }),
+        });
+
+        const { data } = await response.json();
+
+        if (data && data.products && data.products.length) {
+          resolve(true);
+          break;
+        }
+      } catch (err) {
+        console.log(` - Server response error: ${err.message}`);
+      }
+
+      console.log(
+        ` - Rechecking after ${ADMIN_HEALTHCHECK_INTERVAL / 1000} seconds...`
+      );
+      await new Promise((res) => setTimeout(res, ADMIN_HEALTHCHECK_INTERVAL));
+
+      tries--;
+    }
+
+    resolve(false);
+  });
+
 try {
   (async () => {
+    // Check admin if up
+    console.info("Checking admin site...");
+    const isAdminUp = await adminHealthCheck();
+
+    if (!isAdminUp) {
+      console.error(
+        "Admin is still not up for a long time. Skipping prerender."
+      );
+      process.exit();
+    }
+
+    console.info("Admin is up! Proceeding with prerender.");
+
     // Ensure empty out dir
     console.info("Cleaning up output directories.");
     fs.rmdirSync(OUT_DIR, { recursive: true, force: true });
